make it easier to work with single RangeInclusive<T> values
- Added a `From<RangeInclusive<T>>` constructor - Implemented `SortedStarts<T>`
and `SortedDisjoint<T>` for `Option<T>::IntoIter`

These changes make interfacing with single range values easier. It's especially
nice when you want to convert a range into a RangeSetBlaze, or when you want to
perform a set operation (i.e. &) between a RangeSetBlaze and a single Range.

Please let me know if I totally missed existing support for this functionality.
Cheers and thanks for the excellent library!